Extract feature detection package
