+++ api/queries_pr_test.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/MakeNowJust/heredoc"
 	"github.com/cli/cli/v2/internal/ghrepo"
 	"github.com/cli/cli/v2/pkg/httpmock"
 	"github.com/stretchr/testify/assert"
@@ -45,117 +44,6 @@ func TestBranchDeleteRemote(t *testing.T) {
 			if (err != nil) != tt.expectError {
 				t.Fatalf("unexpected result: %v", err)
 			}
-		})
-	}
-}
-
-func Test_determinePullRequestFeatures(t *testing.T) {
-	tests := []struct {
-		name           string
-		hostname       string
-		queryResponse  map[string]string
-		wantPrFeatures pullRequestFeature
-		wantErr        bool
-	}{
-		{
-			name:     "github.com",
-			hostname: "github.com",
-			wantPrFeatures: pullRequestFeature{
-				HasReviewDecision:       true,
-				HasStatusCheckRollup:    true,
-				HasBranchProtectionRule: true,
-			},
-			wantErr: false,
-		},
-		{
-			name:     "GHE empty response",
-			hostname: "git.my.org",
-			queryResponse: map[string]string{
-				`query PullRequest_fields\b`:  `{"data": {}}`,
-				`query PullRequest_fields2\b`: `{"data": {}}`,
-			},
-			wantPrFeatures: pullRequestFeature{
-				HasReviewDecision:       false,
-				HasStatusCheckRollup:    false,
-				HasBranchProtectionRule: false,
-			},
-			wantErr: false,
-		},
-		{
-			name:     "GHE has reviewDecision",
-			hostname: "git.my.org",
-			queryResponse: map[string]string{
-				`query PullRequest_fields\b`: heredoc.Doc(`
-					{ "data": { "PullRequest": { "fields": [
-						{"name": "foo"},
-						{"name": "reviewDecision"}
-					] } } }
-				`),
-				`query PullRequest_fields2\b`: `{"data": {}}`,
-			},
-			wantPrFeatures: pullRequestFeature{
-				HasReviewDecision:       true,
-				HasStatusCheckRollup:    false,
-				HasBranchProtectionRule: false,
-			},
-			wantErr: false,
-		},
-		{
-			name:     "GHE has statusCheckRollup",
-			hostname: "git.my.org",
-			queryResponse: map[string]string{
-				`query PullRequest_fields\b`: heredoc.Doc(`
-					{ "data": { "Commit": { "fields": [
-						{"name": "foo"},
-						{"name": "statusCheckRollup"}
-					] } } }
-				`),
-				`query PullRequest_fields2\b`: `{"data": {}}`,
-			},
-			wantPrFeatures: pullRequestFeature{
-				HasReviewDecision:       false,
-				HasStatusCheckRollup:    true,
-				HasBranchProtectionRule: false,
-			},
-			wantErr: false,
-		},
-		{
-			name:     "GHE has branchProtectionRule",
-			hostname: "git.my.org",
-			queryResponse: map[string]string{
-				`query PullRequest_fields\b`: `{"data": {}}`,
-				`query PullRequest_fields2\b`: heredoc.Doc(`
-					{ "data": { "Ref": { "fields": [
-						{"name": "foo"},
-						{"name": "branchProtectionRule"}
-					] } } }
-				`),
-			},
-			wantPrFeatures: pullRequestFeature{
-				HasReviewDecision:       false,
-				HasStatusCheckRollup:    false,
-				HasBranchProtectionRule: true,
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeHTTP := &httpmock.Registry{}
-			httpClient := NewHTTPClient(ReplaceTripper(fakeHTTP))
-
-			for query, resp := range tt.queryResponse {
-				fakeHTTP.Register(httpmock.GraphQL(query), httpmock.StringResponse(resp))
-			}
-
-			gotPrFeatures, err := determinePullRequestFeatures(httpClient, tt.hostname)
-			if tt.wantErr {
-				assert.Error(t, err)
-				return
-			} else {
-				assert.NoError(t, err)
-			}
-			assert.Equal(t, tt.wantPrFeatures, gotPrFeatures)
 		})
 	}
 }
